pyproject.toml
[project]
name = "simetrik-prueba"
version = "0.1.0"
description = "Prueba técnica Simetrik - Integraciones"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.0",
]

[project.scripts]
prueba = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
